// bisopeto_kinecomap/types.ts
export enum UserType {
    CITIZEN = 'citizen',
    BUSINESS = 'business',
    COLLECTOR = 'collector',
    ADMIN = 'admin'
}
// ...
export interface NotificationItem {
    id: string;
    title: string;
    message: string;
    type: 'info' | 'success' | 'warning' | 'alert';
    time: string;
    date?: string; 
    read: boolean;
    targetUserId: string;
}